introduce hitIterator for going over posting lists
This decodes offsets lazily, thereby avoiding the intermediate garbage
of delta-decoding the complete posting list to get at candidate
matches.

This prepares for a next change where we bail early if there are many
matches. With lazy decoding, we save on delta decoding work if there
are many matches.

In addition, this cleans up the iteration code.

Change-Id: I4a604095ce404eae27c97c40284416901b29f451