Provide less accurate stats on cancel
When the query is canceled (timeout) or when we have enough
interesting matches, consider all remaining documents in the shard
skipped.

If the query is canceled before we start working on the shard, record
the entire shard as skipped. Previously, in this case, we would go
over all of the docs to compute an accurate document skip count, which
is expensive (loading trigrams).

Change-Id: I8361dedde794ea70284c9ecdd108cffa333f42fd