shards: search repos by decreasing shard rank
This will improve quality when there are many results, yielding
results from more important repositories.

In addition, the ordering disambiguates between matches that are
equally good at the shard level, making search results more
reproducible for queries with lots of results.

Limit parallelism to NumCPU, so we avoid searching shards halfway
through; when we have the choice we want to search the more important
shards to the end.

Change-Id: I45d079e94adb96551b19f0de08b8e545ad45ebfb