Drop coversContent logic
If the trigrams cover the entire substring query, there was no need to
check the filecontent that the match was there.

However, with the advent of UTF-8, we still have to look at the file
contents to find size of the match in bytes, and the optimization
avoids a small amount of work anyway, since the query is short by
definition.

This also avoids short queries using different code paths, leading to
more bugs.

Change-Id: I04095ff688f0d17d221c9c6f318cf8f522fa6647