zoekt-sourcegraph-indexserver: Add sourcegraph specific indexserver (#10)
zoekt-indexserver is based on keeping git repos upto date. Now that we will
instead be using archives we need to write a new indexserver which only has two
responsibilities:

* Fetch list of repos and latest commit
* Run zoekt-archive-index for each (repo, commit)
* Remove shards

The last point is pretty great. Previously when you disabled a repo on
sourcegraph we would still attempt to index it and it would still stay in the
zoekt index. Now it will be deleted.