indexserver: Use a priority queue and API batching (#18)
This is a major change to how indexserver works, and an increase in complexity.
Our goal is to reduce indexing lag for large installations (10k+ repositories).
In a large installation the majority of repositories will not need updating.
However, the indexer was a simple for loop which checked the latest commit and
then tried to index it incrementally. This meant we spent a large amount of time
doing network requests (to find out the commit) and not any actual processing.

The first change is to introduce a queue which contains (repo, commit). This
allows us to have a dedicated goroutine which just pops jobs off and indexes
them. This ensures we are always busy and not blocked by the Sourcegraph API.

Additionally the queue is a priority queue. This means if we know a repo has a
new commit it can be processed before a repo which has not had any changes. We
still process queue we suspect haven't changed, just in case the underlying data
has gone. This is also good since it means we don't couple how incremental
updates work in zoekt-archive-index with our indexer.

The second change is to concurrently check what the latest commit is for a
repository. Gitserver can handle many concurrent resolve-rev requests, so we
take advantage of that.