indexserver: Keep shards in trash for 24 hours before removing (#7)
This commits implements functionality allowing us to restore deleted
shards. This is useful when an admin accidentally disables a large number of
repositories temporarily. Instead of having to recompute all the indexes, we
can restore them.

We rely on mtime to track the age of a trashed shard. Additionally we take
care to treat all shards for a repo together, to prevent restoring partial
indexes of a repository.