api: add HasSymbols to Repository
We will use this to decide if an index contains symbols or not. This is needed
since symbol indexing can be present for none or some of the shards. For none
if symbol indexing is disabled (via a feature flag). Partially after a user
has switched on the feature flag and we are still indexing.

Change-Id: Id1c948367d0cf9fae988a94d06936faf9d911add