build: delete older index versions of shards
When we have indexed v16 of a repo, we need to remove v15 and
before. Otherwise zoekt-webserver will serve both v15 and v16 of a shard =>
duplicate results.

We considered adding this process to the indexserver. However, the cleanup
process happens asynchronously which would leave quite a lot of time where
zoekt-webserver would serve up both formats.

Change-Id: I45705de9088719767ff28c23c1c03a64a46d467c