indexserver: Send list of indexed repos when asking for repos to index
Sourcegraph will now take into account rebalancing when deciding which
repositories to index. As such we need to tell it which repositories we have
indexed such that it can tell us to keep repositories that are not yet indexed
by other replicas.

We also move to using a new endpoint targetted at indexing. Previously we used
the generic "repoList" endpoint.

Additionally we remove requesting the list from sourcegraph out of the debug
request code path. This should improve performance of the page for large
installations.

Change-Id: I1e5872c7a525e0a4734885622bd07f4f5acbc66f