zoekt-archive-index: Enforce flowrate on the socket (#35)
flowrate is a Sourcegraph specific feature to limit the amount of traffic we
generate on the networks. The way it is implemented is it keeps conflicting
with upstream zoekt merges. This rewrites it so we instead enforce it in the
DefaultTransport. This should more reliably enforce flowrate as well as
prevent difficult merge conflicts.