shards: search shards in order of name
This is a Sourcegraph specific change. Zoekt creates a slice of shards search,
and starts searching from the front. As such the order of the slice determines
which results are returned before match limits kick in. The slice is created by
iterating over a go map (random) and then sorted by rank. However, Sourcegraph
doesn't use zoekt's ranking functionality, so every time we searched we iterated
over the list of shards in a random order. This lead to page refreshes on
Sourcegraph having different results each time.

We update Zoekt to sort by repo name, which will give us a deterministic order
we search over across requests. Additionally we sort results by repo name in the
graphqlbackend before truncating results so this should lead to the results
being "mostly" deterministic for users (except near the end of the list).

Change-Id: I9e5a297db8af17167180821af26e7951629be880