indexdata: optionally store any section in memory (#38)
We have recently been seeing poor behaviour around the page cache and MMAP in
production. We haven't tracked down why yet, but this adds a workaround for
installations which have enough RAM. Instead of relying on page cache, we copy
in content. We do this by wrapping IndexFile with cachedIndexFile.

cachedIndexFile has a list of cached blocks. If a read call falls within a
cached block, the overlapping part of the block is returned.

This is opt-in behaviour. You opt-in by setting the environment variable
`IN_MEMORY_CONTENT`. Tested by adding some debug logs and running tests with and
without the envvar set.

Part of https://github.com/sourcegraph/sourcegraph/issues/9001