ci: build and push docker images (#42)
This is a port and improvement of our docker images in infrastructure. It will
build and push zoekt-webserver and zoekt-indexserver images on every push to
master. We use GitHub actions to achieve this. It uses the same version format
as we previously did.

The improvement is to share a base image which contains all the zoekt binaries
and dependencies. The different images then just copy out what they need and
do the sourcegraph specific customizations. This helps us share layers in a
cluster as well as reduce the time it takes to build.