docker: put .ctags.d into $HOME
This is an old regression. universal-ctags looks for .ctags.d files either in
$CWD or $HOME. When we switched zoekt to running as the sourcegraph user we
didn't update  the location of .ctags.d to the correct location (was /.ctags.d).
This regression predates the Dockerfile being introduced into this repo.

This caused indexed commits to exclude symbol results for all the additional
languages we support.

Part of https://github.com/sourcegraph/sourcegraph/issues/9756

Change-Id: Ib80c120fda3635513b6f1ffad9de32c6bf72c5b6