indexserver: check incremental in process
Currently we rely on the incremental check in zoekt-archive-index. If we
decide to skip indexing we don't download the archive. However, to move
to shallow cloning and zoekt-git-index we require a clone for
incremental to work. We won't be doing that initially, so we want to do
the incremental check before needing to clone. This commit gets us
closer to that, but by only modifying how we do archive.

Change-Id: I5f1b08fe7e89fb1a1459a54f16890551967a4cdf