indexserver: align flag defaults with dockerfile
This simplifies running the dockerfile. It also makes it easier for
admins to add the debug flag (which has been requested in the past) and
adjust other flags.

Also note that we previously prepended "http://" to
SRC_FRONTEND_INTERNAL. However, that is incorrect since
SRC_FRONTEND_INTERNAL should already be a well formed URL.

Change-Id: I2de460878362d53908598fd89a19810c7b5e5d94