shards: consolidate events before calling scan
We call scan for each fsnotify event. However, multiple events may
happen during a single call to scan, leading to multiple needless calls
to scan. Instead we can consolidate the events from fsnotify so we only
call scan if it isn't running, or only one more time if it is running.

Change-Id: I7b8c3891a202c28357beb8e4bcf74a3c33b3faf0