indexserver: resolve options not revisions
This is a core change to the design of index server. Previously we would
resolve HEAD for all repos and track that in a queue. This doesn't
nicely extend to a multiple branch world. Instead, we now track index
options for all repos (which contain resolved revisions). This neatly
extends into supporting multiple branches.

Change-Id: I2a0c88d829d56f6e47842f557ca10def646188cd