indexserver: timeout if no output for 5m
We have seen index jobs stall. So we added a coarse 1h timeout. This
commit add a faster acting timeout of 5m if the index job has had no
output. I believe this could be faster, but being conservative here. We
still need to fix the underlying cause of stalling, but in the cases
these do happen we will more quickly continue processing the queue.

Change-Id: Ib978b230c39cf327f433466417f9f60cd0e4d3c2