ctags: skip commands that are too large
In the previous commit we assumed commands will be small enough due to
filenames in linux being smaller than 256. However, git repos can be
handcrafted to contain larger names. So instead we ensure we never
overrun the buffer by skipping filenames that will generate a command
that is too large.

Change-Id: Ie4a7279715eda3b91d168b409fd628ce06ce51cb