rpc: do not set timeout if MaxWallTime is specified
Sourcegraph always sets MaxWallTime, so we should let Zoekt respect
that. The MaxWallTime can be greater than 10s, leading to Zoekt
returning at 10s instead of MaxWallTime.

We now only set a timeout on ctx if MaxWallTime is unset (to protect
against accidently not setting in the future). Additionally we update
the default timeout to 20s (same as Sourcegraph).

Change-Id: I42c4968b58482cc3df8d22337a8443b0aec927c9