write: exact allocation for newLinesIndices
Minor change which makes the function slightly slower, but use much less
memory / allocations. A benchmark on a medium sized repo showed this
function was 4% slower, but 60% less memory allocated. This is a minor
part of indexing so is likely not noticeable. However, was a quick
win.

Can also update the algorithm to use bytes.IndexByte. However, that code
is more tricky and this function doesn't contribute much time indexing.

Change-Id: I2e0f037878db08055c1b7d092667a53a0715c0d1