build: always run checktext
Previously we elided the CheckText check if a filename passed the
IgnoreSizeMax test. However, CheckText is still needed to ensure we have
well formed UTF-8 data. There are also some other checks. The builder
depends on these checks. For example with invalid utf-8 data there is an
infinite loop in the index builder.

This commit will always run CheckText. However, it removes the trigram
max check in the case of large files.

Change-Id: I95c1f452ae28ccacadcc299b896ba3c1a5fca17d