indexserver: fetch Options in batches
We suspect the number of concurrent requests we did lead to socket
exhaustion on large vertically scaled instances. So the configuration
endpoint has been extended to take multiple repos and return the
configuration per repo. The responsibility of doing this efficiently is
moved to the frontend. As such we can simplify the implementation here
and remove concurrency.

Change-Id: Id9f51b8b13483cfeb04be5942166d29e5628c304