indexserver: store sourcegraph repo ID in index
This is information we have wanted access to for a long time. Main use
case is more efficient encoding of which repositories to search (int32
is much smaller than the name string). This won't be read, rather we
want to ship this so we start storing the information.

Right now if the repo ID is not set it won't trigger a re-index. IE it
does not affect "incremental" check. The intention is to not trigger a
reindex of all shards. The following release we can enforce it is set,
then the release after that assume it is available.

Change-Id: Idc08b8cc3aaa99a5efed0eed580641f7aeed5df2