improve performance for content searches
for content searches trying to match multiple terms in the same line, we check
whether the matches of the individual terms intersect before calling the
regex engine. If they do, we continue as before, otherwise we skip the
document.

This optimization is useful whenever terms of the query appear often in
the same document but rarely on the same line.