zoekt-webserver: metrics and logs for watchdog
This adds logging around when the state of the watchdog
changes. Additionally we include metrics around the state of the
watchdog. We have noticed in practice the watchdog failing. This is
likely due to an overloaded system. The extra metrics/logging will help
us see how overloaded the system is.

Change-Id: I9e199bcc3644d52f6262b943cc3243b28d41f214