shards: introduce a scheduler abstraction (#74)
This commit introduces a scheduler abstraction for searches to
use. Every search now starts in the "interactive" queue. But after 5s
will move to the "batch" queue. The idea here is to have a simple
approach of allowing long running queries to run without impacting
interactive queries.

We use WeightedSemaphore as the "queue" (internally has a FIFO
queue). Before this commit we just had one WeightedSemaphore with size
#CPUs. In this commit that semaphore becomes the "interactive"
queue. The "batch" queue is a new semaphore with size 1/4 #CPUs. This
doesn't limit the amount of CPU a search can consume, but rather we use
number of CPUs as a guide for the number of concurrent searches that can
be running.

This commit changes a few things, so view the PR for it so you can see it
commit by commit.