indexserver: log shard lifecycle to disk
We are logging when we update or remove a shard to disk. This is a tab
separated values (TSV) file. We will be using this data as input to a
simulator for shard merging. Additionally this data can be used as input
to interesting graphs.

We use two different files to avoid concurrent processes writing to the
same file.

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>
Change-Id: I9c907bc687612db616fb1f896930362608b198c3